[app][uefi] Split memory and block io code into separate files
