+++ lib/uefi/boot_service_provider.h
@@ -17,9 +17,7 @@
 #ifndef __BOOT_SERVICE_PROVIDER_
 #define __BOOT_SERVICE_PROVIDER_
 
-#include "arch/defines.h"
 #include "boot_service.h"
-#include "kernel/vm.h"
 #include "system_table.h"
 
 void setup_boot_service_table(EfiBootService *service);
@@ -122,10 +120,5 @@ struct EFI_LOADED_IMAGE_PROTOCOL {
 };
 
 static constexpr size_t EFI_LOADED_IMAGE_PROTOCOL_REVISION = 0x1000;
-vmm_aspace_t *set_boot_aspace();
-
-void *alloc_page(void *addr, size_t size, size_t align_log2 = PAGE_SIZE_SHIFT);
-void *alloc_page(size_t size, size_t align_log2 = PAGE_SIZE_SHIFT);
-void *identity_map(void *addr, size_t size);
 
 #endif
