+++ lib/uefi/rules.mk
@@ -11,6 +11,8 @@ MODULE_SRCS += \
 	$(LOCAL_DIR)/relocation.cpp \
 	$(LOCAL_DIR)/text_protocol.cpp \
 	$(LOCAL_DIR)/boot_service_provider.cpp \
+	$(LOCAL_DIR)/memory_protocols.cpp \
+	$(LOCAL_DIR)/blockio_protocols.cpp \
 	$(LOCAL_DIR)/runtime_service_provider.cpp \
 	$(LOCAL_DIR)/switch_stack.S \
 	$(LOCAL_DIR)/configuration_table.cpp \
